backend/order_service/tests/test_order_service.py
import pytest
from sqlalchemy.ext.asyncio import AsyncSession
from app.services.order_service import OrderService
from app.schemas.order import OrderCreate

# Note: This is a simplified test example. In a real scenario, you would need to set up
# a test database and mock the external service calls.

def test_order_service_creation():
    """Test that OrderService can be instantiated"""
    # This is a placeholder test
    assert True

if __name__ == "__main__":
    pytest.main([__file__])