backend/order_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
aiohttp==3.9.1
alembic==1.12.1
redis==5.0.1
hiredis==2.2.3
celery==5.3.4
httpx==0.25.0